+++ src/game/player/elements/say/Say.tsx
@@ -3,6 +3,8 @@ import React, {useEffect, useState} from "react";
 import {SayElementProps} from "@player/elements/say/type";
 import {GameState} from "@core/common/game";
 import Sentence from "@player/elements/say/Sentence";
+import {onlyIf} from "@lib/util/data";
+import {useRatio} from "@player/provider/ratio";
 
 export default function Say(
     {
@@ -16,6 +18,7 @@ export default function Say(
     const [isFinished, setIsFinished] = useState(false);
     const {game} = state;
     const [count, setCount] = useState(0);
+    const {ratio} = useRatio();
 
     const handleComplete = () => {
         setIsFinished(true);
@@ -76,26 +79,59 @@ export default function Say(
         <div>
             {sentence.state.display &&
                 (
-                    <div className={
-                        clsx(
-                            "absolute bottom-0 w-[calc(100%-40px)] min-h-[calc(33%-40px)] m-4 bg-white flex flex-col items-start justify-between",
-                            game.config.elementStyles.say.containerClassName,
-                            className
-                        )
-                    } onClick={onElementClick}>
+                    <div
+                        className={
+                            clsx(
+                                "absolute bottom-0 w-[calc(100%-40px)]",
+                                className,
+                                {
+                                    "min-h-[calc(33%-40px)] bg-white flex flex-col items-start justify-between":
+                                        !game.config.elements.text.useAspectScale,
+                                    [game.config.elementStyles.say.containerClassName]: !game.config.elements.text.useAspectScale,
+                                }
+                            )
+                        }
+                        onClick={onElementClick}
+                        style={{
+                            ...onlyIf<React.CSSProperties>(game.config.elements.text.useAspectScale, {
+                                width: game.config.elements.text.width,
+                                height: game.config.elements.text.height,
+                            }),
+                            ...onlyIf<React.CSSProperties>(game.config.app.debug, {
+                                border: "1px solid green",
+                            }),
+                        }}
+                    >
                         <div
-                            className={clsx("rounded-br-md", game.config.elementStyles.say.nameTextClassName)}>
-                            {sentence.config.character?.state.name}
+                            className={clsx(
+                                {
+                                    "bg-white flex flex-col items-start justify-between": game.config.elements.text.useAspectScale,
+                                    [game.config.elementStyles.say.containerClassName]: game.config.elements.text.useAspectScale,
+                                }
+                            )}
+                            style={{
+                                ...onlyIf<React.CSSProperties>(game.config.elements.text.useAspectScale, {
+                                    transform: `scale(${ratio.state.scale})`,
+                                    transformOrigin: "bottom left",
+                                    width: "100%",
+                                    height: "100%",
+                                }),
+                            }}
+                        >
+                            <div
+                                className={clsx("rounded-br-md", game.config.elementStyles.say.nameTextClassName)}>
+                                {sentence.config.character?.state.name}
+                            </div>
+                            <Sentence
+                                sentence={sentence}
+                                gameState={state}
+                                finished={isFinished}
+                                useTypeEffect={useTypeEffect}
+                                onCompleted={handleComplete}
+                                count={count}
+                            />
+                            <div></div>
                         </div>
-                        <Sentence
-                            sentence={sentence}
-                            gameState={state}
-                            finished={isFinished}
-                            useTypeEffect={useTypeEffect}
-                            onCompleted={handleComplete}
-                            count={count}
-                        />
-                        <div></div>
                     </div>
                 )
             }